Update StreamLineRenderer to only use subsegments of a line source which pass through the computable part of the domain (xmin+1 .. xmax-1) This fixes #18. Previously we would try to start streamlines from evenly spaced points between either end of a linesource. When the linesource extended beyond the boundaries of the domain, we would correctly not attempt to integrate from there, but some variables wound up out of sync.
Now, we check to see if the linesource intersects the bounds of the
domain, and if so we only use the subsegment which lies inside the domain
to start streamlines from. This has the added benefit that if a user
specifies a specific number of streamlines, they will always get that many
lines even when their linesource extends outside the domain.