enable lbsim integration on Windows Resolves #11 This re-enables the UI to load LBM input files. The lbsim project can now be built either as a standalone .exe for running on the commandline, or as a DLL. The main project links against the DLL, and as with our other dependencies automatically copies lbsim.dll to its output directory. Note that this is different from how we handle lbsim under Linux, where we statically link it into the final executable.
lbsim.dll only exports the functions necessary for integration with this
project, namely:
  computeDensity, computeVelocity, doCollision, doStreaming,
  treatBoundary, readParameters, initializeFields
If we rework the lbsimwrapper in the future, we should probably also
consider reorganizing lbsim and its exports.

Compositor no longer has any #ifdefs!

Fixed a minor UI bug in Compositor that caused lbsim timesteps to be
printed incorrectly.